+++ src/common/components/app/App.tsx
@@ -2,12 +2,11 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { QueryCache, ReactQueryCacheProvider } from 'react-query';
 import { withRouter, Switch, Route } from 'react-router-dom';
+import OpenLayer from '../../../features/map/OpenLayer';
+import HankeForm from '../../../features/hanke/form/Form';
 import Layout from './Layout';
 import Main from './Main';
-import store from '../store';
-import OpenLayer from './OpenLayer';
-import FormComponent from './form/Form';
-
+import store from './store';
 import './app.scss';
 
 const queryCache = new QueryCache();
@@ -19,7 +18,7 @@ const App: React.FC = () => (
         <Switch>
           <Route exact path="/" render={() => <Main />} />
           <Route exact path="/openlayer" render={() => <OpenLayer />} />
-          <Route exact path="/form" render={() => <FormComponent />} />
+          <Route exact path="/form" render={() => <HankeForm />} />
         </Switch>
       </Layout>
     </ReactQueryCacheProvider>
