--- src/components/app.scss
@@ -1,19 +1,0 @@
-@font-face {
-  font-family: HelsinkiGrotesk;
-  src: url('../utils/desktop_fonts/HelsinkiGrotesk-Regular.otf') format('opentype');
-}
-
-@font-face {
-  font-family: HelsinkiGrotesk;
-  font-weight: 500;
-  src: url('../utils/desktop_fonts/HelsinkiGrotesk-Medium.otf') format('opentype');
-}
-@font-face {
-  font-family: HelsinkiGrotesk;
-  font-weight: bold;
-  src: url('../utils/desktop_fonts/HelsinkiGrotesk-Bold.otf') format('opentype');
-}
-
-body {
-  font-family: 'HelsinkiGrotesk', serif;
-}
