+++ src/features/hanke/Hanke.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { render, waitForElementToBeRemoved } from '@testing-library/react';
-import Project from './Project';
+import Hanke from './Hanke';
 
 test('renders learn react link', async () => {
-  const { getByText } = render(<Project />);
+  const { getByText } = render(<Hanke />);
 
   await waitForElementToBeRemoved(() => getByText('Ladataan...'));
