+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
-import App from './components/App';
-import './i18n/i18n';
+import App from './common/components/app/App';
+import './locales/i18n';
 
 // import * as serviceWorker from './serviceWorker';
